LayoutTests/imported/w3c/web-platform-tests/storage-access-api/requestStorageAccess-cross-origin-iframe-navigation.sub.https.window-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: SyntaxError: The string did not match the expected pattern.

Harness Error (FAIL), message = Unhandled rejection: The string did not match the expected pattern.

TIMEOUT Self-initiated reloads preserve storage access Test timed out
NOTRUN Self-initiated same-origin navigations preserve storage access
NOTRUN Non-self-initiated same-origin navigations do not preserve storage access
NOTRUN Self-initiated cross-origin navigations do not preserve storage access
